fix(server): recover retries and bound callback cycles
